+++ extensions/simplebridge/driver.go
@@ -1,20 +1,13 @@
 package simplebridge
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
-	"net"
-	"strconv"
 	"sync"
 
-	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/network"
 	"github.com/docker/docker/sandbox"
 	"github.com/docker/docker/state"
-
-	"github.com/vishvananda/netlink"
-	"github.com/vishvananda/netlink/nl"
 )
 
 const (
@@ -36,61 +29,6 @@ func (d *BridgeDriver) GetNetwork(network string) (network.Network, error) {
 func (d *BridgeDriver) Restore(s state.State) error {
 	d.state = s
 	return nil
-}
-
-func (d *BridgeDriver) loadEndpoint(network, endpoint string) (*BridgeEndpoint, error) {
-	scope := d.schema.Endpoint(network, endpoint)
-
-	iface, err := scope.Get("interface_name")
-	if err != nil {
-		return nil, err
-	}
-
-	hwAddr, err := scope.Get("hwaddr")
-	if err != nil {
-		return nil, err
-	}
-
-	mtu, err := scope.Get("mtu")
-	if err != nil {
-		return nil, err
-	}
-
-	ipaddr, err := scope.Get("ip")
-	if err != nil {
-		return nil, err
-	}
-
-	ip := net.ParseIP(ipaddr)
-
-	mtuInt, _ := strconv.ParseUint(mtu, 10, 32)
-
-	netObj, err := d.loadNetwork(network)
-	if err != nil {
-		return nil, err
-	}
-
-	return &BridgeEndpoint{
-		ID:            endpoint,
-		interfaceName: iface,
-		hwAddr:        hwAddr,
-		mtu:           uint(mtuInt),
-		network:       netObj,
-		ip:            ip,
-	}, nil
-}
-
-func (d *BridgeDriver) saveEndpoint(network string, ep *BridgeEndpoint) error {
-	scope := d.schema.Endpoint(network, ep.ID)
-
-	pathMap := map[string]string{
-		"interface_name": ep.interfaceName,
-		"hwaddr":         ep.hwAddr,
-		"mtu":            strconv.Itoa(int(ep.mtu)),
-		"ip":             ep.ip.String(),
-	}
-
-	return scope.MultiSet(pathMap)
 }
 
 // discovery driver? should it be hooked here or in the core?
@@ -154,57 +92,6 @@ func (d *BridgeDriver) Unlink(netid, name string, sb sandbox.Sandbox) error {
 	return nil
 }
 
-func (d *BridgeDriver) saveNetwork(network string, bridge *BridgeNetwork) error {
-	networkSchema := d.schema.Network(network)
-	// FIXME allocator, address will be broken if not saved
-	if err := networkSchema.Set("bridge_interface", bridge.bridge.Name); err != nil {
-		return err
-	}
-
-	if err := networkSchema.Set("address", bridge.network.String()); err != nil {
-		return err
-	}
-
-	if bridge.vxlan != nil {
-		networkSchema.Set("vxlan_device", bridge.vxlan.Attrs().Name)
-	}
-
-	return nil
-}
-
-func (d *BridgeDriver) loadNetwork(network string) (*BridgeNetwork, error) {
-	networkSchema := d.schema.Network(network)
-
-	iface, err := networkSchema.Get("bridge_interface")
-	if err != nil {
-		return nil, err
-	}
-
-	addr, err := networkSchema.Get("address")
-	if err != nil {
-		return nil, err
-	}
-
-	ip, ipNet, err := net.ParseCIDR(addr)
-	ipNet.IP = ip
-
-	var vxlan *netlink.Vxlan
-
-	vxdev, err := networkSchema.Get("vxlan_device")
-	if err == nil && vxdev != "" {
-		vxlan = &netlink.Vxlan{LinkAttrs: netlink.LinkAttrs{Name: vxdev}}
-	}
-
-	return &BridgeNetwork{
-		vxlan:       vxlan,
-		bridge:      &netlink.Bridge{LinkAttrs: netlink.LinkAttrs{Name: iface}},
-		ID:          network,
-		driver:      d,
-		network:     ipNet,
-		ipallocator: NewIPAllocator(iface, ipNet, nil, nil),
-	}, nil
-}
-
 func (d *BridgeDriver) AddNetwork(network string, args []string) error {
 	// FIXME this should be abstracted from the network driver
 
@@ -247,108 +134,4 @@ func (d *BridgeDriver) RemoveNetwork(network string) error {
 	}
 
 	return bridge.destroy()
-}
-
-func (d *BridgeDriver) createBridge(id string, vlanid uint, port uint, peer, device string) (*BridgeNetwork, error) {
-	dockerbridge := &netlink.Bridge{netlink.LinkAttrs{Name: id}}
-
-	linkval, err := d.getInterface(id, dockerbridge)
-	if err != nil {
-		log.Println("Error get interface", err)
-		return nil, err
-	}
-	dockerbridge = linkval.(*netlink.Bridge)
-
-	addr, err := GetBridgeIP()
-	if err != nil {
-		return nil, err
-	}
-
-	addrList, err := netlink.AddrList(dockerbridge, nl.GetIPFamily(addr.IP))
-	if err != nil {
-		return nil, err
-	}
-
-	var found bool
-	for _, el := range addrList {
-		if bytes.Equal(el.IPNet.IP, addr.IP) && bytes.Equal(el.IPNet.Mask, addr.Mask) {
-			found = true
-			break
-		}
-	}
-	if !found {
-		if err := netlink.AddrAdd(dockerbridge, &netlink.Addr{IPNet: addr}); err != nil {
-			log.Println("Error add addr", err)
-			return nil, err
-		}
-	}
-
-	if err := netlink.LinkSetUp(dockerbridge); err != nil {
-		log.Println("Error up bridge", err)
-		return nil, err
-	}
-
-	if err := setupIPTables(id, addr); err != nil {
-		return nil, err
-	}
-
-	var vxlan *netlink.Vxlan
-
-	if peer != "" && device != "" {
-		iface, err := net.InterfaceByName(device)
-		if err != nil {
-			log.Println("Error get interface", err)
-			return nil, err
-		}
-
-		vxlan = &netlink.Vxlan{
-			// DEMO FIXME: name collisions, better error recovery
-			LinkAttrs:    netlink.LinkAttrs{Name: "vx" + id, Flags: net.FlagMulticast},
-			VtepDevIndex: iface.Index,
-			VxlanId:      int(vlanid),
-			Group:        net.ParseIP(peer),
-			Port:         int(port),
-		}
-
-		linkval, err = d.getInterface(vxlan.LinkAttrs.Name, vxlan)
-		if err != nil {
-			log.Println("Error get interface", err)
-			return nil, err
-		}
-		vxlan = linkval.(*netlink.Vxlan)
-
-		// ignore errors in case it was already set
-		if err := netlink.LinkSetMaster(vxlan, dockerbridge); err != nil {
-			log.Println("Error linksetmaster", err)
-			return nil, err
-		}
-		if err := netlink.LinkSetUp(vxlan); err != nil {
-			log.Println("Error linksetmaster", err)
-			return nil, err
-		}
-	}
-
-	if err := MakeChain(id, dockerbridge.LinkAttrs.Name); err != nil {
-		return nil, err
-	}
-
-	return &BridgeNetwork{
-		vxlan:       vxlan,
-		bridge:      dockerbridge,
-		ID:          id,
-		driver:      d,
-		network:     addr,
-		ipallocator: NewIPAllocator(dockerbridge.LinkAttrs.Name, addr, nil, nil),
-	}, nil
-}
-
-func (d *BridgeDriver) destroyBridge(b *netlink.Bridge, v *netlink.Vxlan) error {
-	// DEMO FIXME
-	if v != nil {
-		if err := netlink.LinkDel(v); err != nil {
-			return err
-		}
-	}
-
-	return netlink.LinkDel(b)
 }
